requirements.txt
# Core ML Framework
torch==2.2.0
torchvision==0.17.0

# Web Framework for Server
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Client for Communication
requests==2.31.0

# Configuration Management
PyYAML==6.0.1
pydantic==2.5.0

# System Metrics Collection
psutil==5.9.6

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Database
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Database
psycopg2-binary==2.9.7

# SSH Connectivity for Raspberry Pi management
paramiko==3.4.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
